+++ merkle-tree/src/hybrid/hybrid_strategy/node_converter.rs
@@ -1,0 +1,7 @@
+pub struct NodeConversor256BabyBearBytes {}
+
+impl NodeConverter<[BabyBear; 8], [u8; 32]> for NodeConversor256BabyBearBytes {
+    fn to_n1(&self, input: [BabyBear; 8]) -> [u8; 32] {
+        todo!()
+    }
+}
